+++ ecs/EcsLib/RegistrySystem.h
@@ -10,7 +10,7 @@ class RegistrySystem : public System<RegistrySystem> {
 protected:
     bool Initialize() override;
     void Start() override;
-    void Update(float dt) override;
+    void Update(double dt) override;
     void Stop() override;
     void Uninitialize() override;
